refactor(bridge): extract message action builders
